+++ neo/idlib/math/VecX.cpp
@@ -37,7 +37,7 @@ If you have questions concerning this license or the applicable additional terms
 
 float	idVecX::temp[VECX_MAX_TEMP+4];
 float *	idVecX::tempPtr = reinterpret_cast<float*>((reinterpret_cast<UINT_PTR>(idVecX::temp) + 15) & ~15);
-int		idVecX::tempIndex = 0;
+int64	idVecX::tempIndex = 0;
 
 /*
 =============
