+++ neo/idlib/sys/sys_filesystem.h
@@ -45,7 +45,7 @@ sysFolder_t		Sys_IsFolder( const char *path );
 
 // use fs_debug to verbose Sys_ListFiles
 // returns -1 if directory was not found (the list is cleared)
-int				Sys_ListFiles( const char * directory, const char * extension, idList<class idStr> & list );
+int64			Sys_ListFiles( const char * directory, const char * extension, idList<class idStr> & list );
 
 const char *	Sys_EXEPath();
 const char *	Sys_CWD();
